+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.0.2-nullsafety.2] - 3/4/2021
+
+* introducing Captain API to RouteInformationParser
+
+## [0.0.2-nullsafety.1] - 2/12/2021
+
+* another try with stable 2.12-0-0
+
 ## [0.0.2-nullsafety] - 2/12/2021
 
 * introduced null-safety + made popPage, shouldPop, actions optional/nullable